+++ src/SampleApp/Form1.Designer.cs
@@ -1,3 +1,5 @@
+using L10NSharp.XLiffUtils;
+
 namespace SampleApp
 {
     partial class Form1
@@ -43,14 +45,14 @@ namespace SampleApp
 			((System.ComponentModel.ISupportInitialize)(this.localizationExtender1)).BeginInit();
 			((System.ComponentModel.ISupportInitialize)(this.dataGridView1)).BeginInit();
 			this.SuspendLayout();
-			// 
+			//
 			// localizationExtender1
-			// 
+			//
 			this.localizationExtender1.LocalizationManagerId = "SampleApp";
 			this.localizationExtender1.PrefixForNewItems = "TheSampleForm";
-			// 
+			//
 			// _getDynamicStringButton
-			// 
+			//
 			this.localizationExtender1.SetLocalizableToolTip(this._getDynamicStringButton, null);
 			this.localizationExtender1.SetLocalizationComment(this._getDynamicStringButton, null);
 			this.localizationExtender1.SetLocalizingId(this._getDynamicStringButton, "TheSampleForm.button1");
@@ -61,9 +63,9 @@ namespace SampleApp
 			this._getDynamicStringButton.Text = "Get Name Dynamically";
 			this._getDynamicStringButton.UseVisualStyleBackColor = true;
 			this._getDynamicStringButton.Click += new System.EventHandler(this.button1_Click);
-			// 
+			//
 			// label1
-			// 
+			//
 			this.label1.AutoSize = true;
 			this.localizationExtender1.SetLocalizableToolTip(this.label1, null);
 			this.localizationExtender1.SetLocalizationComment(this.label1, null);
@@ -73,9 +75,9 @@ namespace SampleApp
 			this.label1.Size = new System.Drawing.Size(73, 13);
 			this.label1.TabIndex = 1;
 			this.label1.Text = "A Static Label";
-			// 
+			//
 			// label2
-			// 
+			//
 			this.label2.AutoSize = true;
 			this.localizationExtender1.SetLocalizableToolTip(this.label2, null);
 			this.localizationExtender1.SetLocalizationComment(this.label2, null);
@@ -85,9 +87,9 @@ namespace SampleApp
 			this.label2.Size = new System.Drawing.Size(202, 13);
 			this.label2.TabIndex = 2;
 			this.label2.Text = "or Alt+Shift Click on items to localize them";
-			// 
+			//
 			// uiLanguageComboBox1
-			// 
+			//
 			this.uiLanguageComboBox1.DisplayMember = "NativeName";
 			this.uiLanguageComboBox1.DropDownHeight = 200;
 			this.uiLanguageComboBox1.DropDownStyle = System.Windows.Forms.ComboBoxStyle.DropDownList;
@@ -103,9 +105,9 @@ namespace SampleApp
 			this.uiLanguageComboBox1.Size = new System.Drawing.Size(121, 23);
 			this.uiLanguageComboBox1.TabIndex = 3;
 			this.uiLanguageComboBox1.SelectedIndexChanged += new System.EventHandler(this.uiLanguageComboBox1_SelectedIndexChanged);
-			// 
+			//
 			// linkLabel1
-			// 
+			//
 			this.linkLabel1.AutoSize = true;
 			this.localizationExtender1.SetLocalizableToolTip(this.linkLabel1, null);
 			this.localizationExtender1.SetLocalizationComment(this.linkLabel1, null);
@@ -117,23 +119,23 @@ namespace SampleApp
 			this.linkLabel1.TabStop = true;
 			this.linkLabel1.Text = "Open Translation Dialog,";
 			this.linkLabel1.LinkClicked += new System.Windows.Forms.LinkLabelLinkClickedEventHandler(this.linkLabel1_LinkClicked);
-			// 
+			//
 			// columnHeader1
-			// 
+			//
 			this.localizationExtender1.SetLocalizableToolTip(this.columnHeader1, null);
 			this.localizationExtender1.SetLocalizationComment(this.columnHeader1, null);
 			this.localizationExtender1.SetLocalizingId(this.columnHeader1, "TheSampleForm.columnHeader1");
 			this.columnHeader1.Text = "One";
-			// 
+			//
 			// columnHeader2
-			// 
+			//
 			this.localizationExtender1.SetLocalizableToolTip(this.columnHeader2, null);
 			this.localizationExtender1.SetLocalizationComment(this.columnHeader2, null);
 			this.localizationExtender1.SetLocalizingId(this.columnHeader2, "TheSampleForm.columnHeader2");
 			this.columnHeader2.Text = "Two";
-			// 
+			//
 			// dataGridView1
-			// 
+			//
 			this.dataGridView1.ColumnHeadersHeightSizeMode = System.Windows.Forms.DataGridViewColumnHeadersHeightSizeMode.AutoSize;
 			this.dataGridView1.Columns.AddRange(new System.Windows.Forms.DataGridViewColumn[] {
             this.Column1});
@@ -144,9 +146,9 @@ namespace SampleApp
 			this.dataGridView1.Name = "dataGridView1";
 			this.dataGridView1.Size = new System.Drawing.Size(446, 73);
 			this.dataGridView1.TabIndex = 6;
-			// 
+			//
 			// listView1
-			// 
+			//
 			this.listView1.Columns.AddRange(new System.Windows.Forms.ColumnHeader[] {
             this.columnHeader1,
             this.columnHeader2});
@@ -156,16 +158,16 @@ namespace SampleApp
 			this.listView1.TabIndex = 5;
 			this.listView1.UseCompatibleStateImageBehavior = false;
 			this.listView1.View = System.Windows.Forms.View.Details;
-			// 
+			//
 			// Column1
-			// 
+			//
 			this.Column1.AutoSizeMode = System.Windows.Forms.DataGridViewAutoSizeColumnMode.Fill;
 			this.Column1.HeaderText = "_L10N_:TheSampleForm.SampleDataGridView.ColumnHeadings.FirstColumn!First";
 			this.Column1.Name = "Column1";
 			this.Column1.ToolTipText = "My tooltip";
-			// 
+			//
 			// Form1
-			// 
+			//
 			this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
 			this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
 			this.ClientSize = new System.Drawing.Size(559, 462);