+++ CHANGELOG.md
@@ -16,6 +16,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- Allow to select translation memory (TMX or XLIFF). This changed a few APIs.
+  To create a `LocalizationManager` you now pass a `TranslationMemory` parameter:
+  
+      LocalizationManager.Create(TranslationMemory.XLiff, lang, "SampleApp", "SampleApp",
+        Application.ProductVersion, directoryOfInstalledXliffFiles, "MyCompany/L10NSharpSample",
+        icon, "sample@example.com", "SampleApp");
+
 ## [3.1.1] - 2019-04-26
 
 ### Fixed